structure changes type moved to dedicated file
