+++ components/HomeHero/HeroCarousel/HeroCarouselHelpers.ts
