+++ sanity/client.ts
@@ -4,10 +4,6 @@ const projectId = process.env.NEXT_PUBLIC_SANITY_PROJECT_ID
 const dataset = process.env.NEXT_PUBLIC_SANITY_DATASET
 const apiVersion = process.env.NEXT_PUBLIC_SANITY_API_VERSION || 'v2023-11-21'
 
-console.log('projectId:', projectId);
-console.log('dataset:', dataset);
-console.log('apiVersion:', apiVersion);
-
 export const client = createClient({
   projectId,
   dataset,
